@@ -5,39 +5,59 @@
 	"description": "Inject CSS styles on demand for ReactJS Components. Keeps your DOM lean.",
 	"license": "BSD-3-Clause",
 	"dependencies": {
+		"@babel/runtime-corejs2": "^7.6.3",
 		"@bitchcraft/unicorn-logger": "^1.1.1",
-		"recompose": "^0.26.0",
-		"uuid": "^3.2.1"
+		"recompose": "^0.30.0",
+		"uuid": "^3.3.3"
 	},
 	"devDependencies": {
-		"babel-cli": "^6.18.0",
-		"babel-core": "^6.20.0",
-		"babel-eslint": "^8.2.2",
-		"babel-jest": "^20.0.3",
-		"babel-loader": "^7.1.3",
-		"babel-plugin-add-module-exports": "^0.2.1",
-		"babel-plugin-module-resolver": "^3.1.0",
-		"babel-plugin-transform-runtime": "^6.15.0",
-		"babel-preset-env": "^1.6.1",
-		"babel-preset-flow": "^6.23.0",
-		"babel-preset-react": "^6.24.1",
-		"babel-preset-stage-0": "^6.16.0",
-		"eslint": "^4.18.1",
-		"eslint-config-airbnb-base": "^12.1.0",
-		"eslint-plugin-flowtype": "^2.46.1",
-		"eslint-plugin-flowtype-errors": "^3.3.7",
-		"eslint-plugin-import": "^2.9.0",
-		"expect": "^1.13.2",
+		"@babel/cli": "^7.0.0",
+		"@babel/core": "^7.0.0",
+		"@babel/plugin-proposal-class-properties": "^7.0.0",
+		"@babel/plugin-proposal-decorators": "^7.0.0",
+		"@babel/plugin-proposal-do-expressions": "^7.0.0",
+		"@babel/plugin-proposal-export-default-from": "^7.0.0",
+		"@babel/plugin-proposal-export-namespace-from": "^7.0.0",
+		"@babel/plugin-proposal-function-bind": "^7.0.0",
+		"@babel/plugin-proposal-function-sent": "^7.0.0",
+		"@babel/plugin-proposal-json-strings": "^7.0.0",
+		"@babel/plugin-proposal-logical-assignment-operators": "^7.0.0",
+		"@babel/plugin-proposal-nullish-coalescing-operator": "^7.0.0",
+		"@babel/plugin-proposal-numeric-separator": "^7.0.0",
+		"@babel/plugin-proposal-optional-chaining": "^7.0.0",
+		"@babel/plugin-proposal-pipeline-operator": "^7.0.0",
+		"@babel/plugin-proposal-throw-expressions": "^7.0.0",
+		"@babel/plugin-syntax-dynamic-import": "^7.0.0",
+		"@babel/plugin-syntax-import-meta": "^7.0.0",
+		"@babel/plugin-transform-runtime": "^7.0.0",
+		"@babel/preset-env": "^7.0.0",
+		"@babel/preset-flow": "^7.0.0",
+		"@babel/preset-react": "^7.0.0",
+		"babel-core": "^7.0.0-bridge.0",
+		"babel-eslint": "^10.0.3",
+		"babel-jest": "^24.9.0",
+		"babel-loader": "^8.0.6",
+		"babel-plugin-add-module-exports": "^1.0.2",
+		"babel-plugin-module-resolver": "^3.2.0",
+		"eslint": "^6.5.1",
+		"eslint-config-airbnb": "^18.0.1",
+		"eslint-formatter-pretty": "^2.1.1",
+		"eslint-plugin-flowtype": "^4.3.0",
+		"eslint-plugin-flowtype-errors": "^4.1.0",
+		"eslint-plugin-import": "^2.18.2",
+		"eslint-plugin-jsx-a11y": "^6.2.3",
+		"eslint-plugin-react": "^7.16.0",
+		"expect": "^24.9.0",
 		"extract-text-webpack-plugin": "^3.0.2",
-		"file-loader": "^1.1.11",
-		"flow-bin": "^0.66.0",
-		"jest": "^20.0.4",
-		"jsdom": "^11.6.2",
-		"react": "^16.2.0",
-		"react-test-renderer": "^16.2.0",
-		"remark-preset-lint-recommended": "^3.0.1",
-		"webpack": ">=1.14.0 < 5.0.0",
-		"webpack-cli": "^3.3.2"
+		"file-loader": "^4.2.0",
+		"flow-bin": "^0.109.0",
+		"jest": "^24.9.0",
+		"jsdom": "^15.1.1",
+		"react": "^16.10.2",
+		"react-test-renderer": "^16.10.2",
+		"remark-preset-lint-recommended": "^3.0.3",
+		"webpack": "^4.41.0",
+		"webpack-cli": "^3.3.9"
 	},
 	"peerDependencies": {
 		"autoprefixer": "^8.1.0",
@@ -64,6 +84,7 @@
 	"scripts": {
 		"test": "jest",
 		"prestart": "yarn install && yarn run test src",
+		"lint": "eslint --cache --format=eslint-formatter-pretty .",
 		"start": "yarn run bundle",
 		"bundle": "NODE_ENV=production webpack --progress && yarn run test lib"
 	},
