@@ -1,13 +1,13 @@
 /* eslint-env jest */
 /* eslint-disable no-unused-vars */
 import React from 'react';
-import withInjector from './index';
 import renderer from 'react-test-renderer';
+import withInjector from './index';
 
 const MyComponent = ({ children }) => React.createElement('span', { children });
 MyComponent.displayName = 'MyComponent';
-const stylesheet = values => ({});
-const style = theme => ({});
+const stylesheet = (values) => ({});
+const style = (theme) => ({});
 const MyWrappedComponent = withInjector(stylesheet, style)(MyComponent);
 
 test('inject()', () => {
